pyproject.toml
[project]
name = "report-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.11",
    "google-generativeai>=0.8.4",
    "gradio>=5.20.0",
    "ipykernel>=6.29.5",
    "jupyter>=1.1.1",
    "openai>=1.65.2",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "unidecode>=1.3.8",
    "uvicorn>=0.34.0",
]
